requirements.txt
celery==5.0.1
django-celery-beat==2.1.0
Django==3.1.5
gunicorn==20.0.4
psycopg2==2.8.6
redis==3.5.3
kafka-python==2.0.1